components/modules/analysis_utils/pyproject.toml
[tool.poetry]
name = "analysis_utils"
version = "0.1.0"
description = ""
authors = ["John Shanahan <shanahan.jrs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
msgpack = "^0.6.1"
pydantic = "^1.6.1"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.dephell.main]
envs = ["main"]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

